backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic==0.40.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.4
pillow==11.1.0
aiofiles==24.1.0
openai>=1.50.0
PyYAML>=6.0
